+++ config/mapprovider/yamlmapprovider/mapprovider.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package configmapprovider // import "go.opentelemetry.io/collector/config/configmapprovider"
+package yamlmapprovider // import "go.opentelemetry.io/collector/config/mapprovider/yamlmapprovider"
 
 import (
 	"context"
@@ -24,11 +24,11 @@ import (
 	"go.opentelemetry.io/collector/config"
 )
 
-const bytesSchemeName = "yaml"
+const schemeName = "yaml"
 
-type yamlMapProvider struct{}
+type mapProvider struct{}
 
-// NewYAML returns a new Provider that allows to provide yaml bytes.
+// New returns a new config.MapProvider that allows to provide yaml bytes.
 //
 // This Provider supports "yaml" scheme, and can be called with a "location" that follows:
 //   bytes-location   = "yaml:" yaml-bytes
@@ -36,23 +36,23 @@ type yamlMapProvider struct{}
 // Examples:
 // `yaml:processors::batch::timeout: 2s`
 // `yaml:processors::batch/foo::timeout: 3s`
-func NewYAML() Provider {
-	return &yamlMapProvider{}
+func New() config.MapProvider {
+	return &mapProvider{}
 }
 
-func (s *yamlMapProvider) Retrieve(_ context.Context, location string, _ WatcherFunc) (Retrieved, error) {
-	if !strings.HasPrefix(location, bytesSchemeName+":") {
-		return Retrieved{}, fmt.Errorf("%v location is not supported by %v provider", location, bytesSchemeName)
+func (s *mapProvider) Retrieve(_ context.Context, location string, _ config.WatcherFunc) (config.Retrieved, error) {
+	if !strings.HasPrefix(location, schemeName+":") {
+		return config.Retrieved{}, fmt.Errorf("%v location is not supported by %v provider", location, schemeName)
 	}
 
 	var data map[string]interface{}
-	if err := yaml.Unmarshal([]byte(location[len(bytesSchemeName)+1:]), &data); err != nil {
-		return Retrieved{}, fmt.Errorf("unable to parse yaml: %w", err)
+	if err := yaml.Unmarshal([]byte(location[len(schemeName)+1:]), &data); err != nil {
+		return config.Retrieved{}, fmt.Errorf("unable to parse yaml: %w", err)
 	}
 
-	return Retrieved{Map: config.NewMapFromStringMap(data)}, nil
+	return config.Retrieved{Map: config.NewMapFromStringMap(data)}, nil
 }
 
-func (s *yamlMapProvider) Shutdown(context.Context) error {
+func (s *mapProvider) Shutdown(context.Context) error {
 	return nil
 }